make threaded exit and cache file loading safer by pausing other threads
Also add a missing lock call in jl_safepoint_suspend_thread.

Instead of only preventing access to LLVM and inference (via the codegen
lock), this prevents further access to all internals by disabling the
safepoints and waiting for the threads to stop (or to be in unmanaged
code that we don't care about).

This then uses that to also make staticdata loading a bit safer, since
`jl_lookup_cache_type_` assumes there are no concurrent mutator threads,
and we can make that true by pausing all of those.