make codegen threadsafe, sinking the necessary lock now into JuliaOJIT (#55106)
This adds a new helper `jl_read_codeinst_invoke` that should help manage
reading the state out of a CodeInstance correctly everywhere. Then
replaces all of the places where we have optimizations in codegen where
we check for this (to build a name in the JIT for it) with that call.
And finally moves the `jl_codegen_lock` into
`jl_ExecutionEngine->jitlock` so that it is now more clear that this is
only protecting concurrent access to the JIT state it manages (which
includes the invoke field of all CodeInstance objects). In a subsequent
followup, that `jitlock` and `codeinst_in_flight` will be replaced with
something akin to the new engine (for CodeInfo inference) which helps
partition that JIT lock mechanism (for CodeInstance / JIT insertion) to
correspond just to a single CodeInstance, and not globally to all of
them.