Simplify sweeping of big values (#54936)
Simplifies the layout of the doubly linked list of big objects to make
it a bit more canonical: let's just store a pointer to the previous
element, instead of storing a "pointer to the next element of the
previous element". This should make the implementation a bit easier to
understand without incurring any memory overhead.

I ran the serial and multithreaded benchmarks from GCBenchmarks and this
seems fairly close to performance neutral on my machine. We also ran our
internal benchmarks on it at RAI and it looks fine from a correctness
and performance point of view.

---------

Co-authored-by: Kiran Pamnany <kpamnany@users.noreply.github.com>