Improve static compilabiliaty of Printf
This makes small tweaks to "compiler hints" in Printf to reduce the
number of runtime dispatches. For example, on `master` calling

```
prnt(a::AbstractString, b::AbstractFloat, c::Union{AbstractChar,AbstractString}) =
    @printf("%5s    %12.5e    %c\n", a, b, c)
```

results in 4 runtime-dispatches, whereas on this branch there is just 1
(which comes from not knowing the type of `Main.stdout`).

This also removes a number of manual `@inline` annotations. These are on
fairly large `fmt` methods, many of which have loops; forcing inlining
seems likely to bloat the code without much benefit.