Allow `Kind`s to be registered by packages outside JuliaSyntax (JuliaLang/JuliaSyntax.jl#461)
Extensible kinds are quite tricky. We want
* To use a small number of bits for them
* To have the string representation in the source, but have the compiler
  able to fully inline the integer representation.
* Allow modules with different kinds to cooperate together on the same
  integer representation.
* Not trigger invalidation when new kinds are added
* Different `Kind` modules to not require cooperation

This is a very hard set of constraints to satisfy. The last one is
already impossible in a single flat namespace so in this design we've
given up on it and require cooperation between all kind extension
modules, including module authors allocating non-colliding id's for
their modules, in addition to non-colliding kind names.