loading: Don't assume a package is loaded by its own name
Recent discussion on slack has revealed that there is some disagreement
about what it means for a package to have a different name in the
`deps` section for one of its dependencies than said dependency itself
declares in its own Project.toml.

There's basically two views:
1. This situation shouldn't arise, the primary purpose of the deps section is
   to disambiguate between two packages of the same name (but with different
   UUIDs).

2. The `Project.toml` file declares a mapping `name` => `uuid` that all loading
   operations within said package operate on and the actual name of the package
   is largely irrelevant.

Worse, the current code doesn't really do either. It mostly behaves like the
second, except that it doesn't actually work, because it forms the default
entrypath using the name from the deps section, not the Project.toml (resp.
a copy of it in the Manifest.toml). As a result, the only time you're allowed
to use a name other than what the project declares in a `[deps]` section
is when the dependency explicitly declares an `entrypath` (which is a relatively
new feature).

This PR adjusts loading to fully behave like the second interpretation.
For example, having the following in

Project.toml
```
[deps]
GeneralHTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
```

```
[[deps.GeneralHTTP]]
name = "HTTP"
deps = ["Base64", "CodecZlib", "ConcurrentUtilities", "Dates", "ExceptionUnwrapping", "Logging", "LoggingExtras", "MbedTLS", "NetworkOptions", "OpenSSL", "Random", "SimpleBufferStream", "Sockets", "URIs", "UUIDs"]
git-tree-sha1 = "d1d712be3164d61d1fb98e7ce9bcbc6cc06b45ed"
uuid = "cd3eb016-35fb-5094-929b-558a96fad6f3"
version = "1.10.8"
```

would allow `using GeneralHTTP` to load the package otherwise known
as `HTTP`.

This is only the loading side of this. While Pkg seems surprisingly
ok with this also, it is definitely not tested and likely to cause
confusing behavior, so the primary purpose of this PR is to decide
which behavior we think is correct and then to gradually bring the
implementation into compliance.