inference: Allow any non-nothing source in ci_has_source for extern owner
In updating DAECompiler to the new engine API, I was running some
trouble, because the source it produces is not a `CodeInfo` or a
`String` (which is fine because it has a non-nothing `owner`).
Adjust the test to consider any non-nothing value in .inferred valid
source (at least for non-nothing CI owners).