Ensure bidiagonal setindex! does not read indices in error message (#55342)
This fixes the error message if the matrix is uninitialized. This is
because a `Bidiagonal` with `uplo == 'L'` may still be `istriu` if the
subdiaognal is zero. We only care about the band index in the error
message, and not the values.

(cherry picked from commit f2f188d57346a0163b82740ac0a758311c41004f)