Add functions to manipulate the GC heap counters
`jl_malloc` and `jl_free` can be used to inform the GC about memory
pressure caused by external libraries. This PR adds two functions
that only do the accounting and not the memory allocation/freeing.

The intended audience is users of libraries that do their own memory
management, but still have establishes sizes and deallocation points.
An example of this is the unified memory API in CUDA.