Address atomicness concerns
Current julia master changes binding kinds implicitly based
on the status of the owner field and does so atomically.
The fully partitioned version of this will do so using atomic
operations on the partition list.

This current state is a bit of a weird hybrid where we want
to change both the ->kind and the ->restriction field. To allow
this to happen, merge them into one field that has custom GC
support to ignore the low bits.