effects: minor fixes for the effects system correctness
This commit implements several fixes related to the correctness of the
effect system. The most significant change addresses an issue where
post-opt analysis was not correctly propagating the taints of `:noub`
and `:nortcall` of `:foreigncall` expressions, which could lead to
incorrect effect bits. Additionally, adjustments have been made to
the values of effects used in various worst-case scenarios.