re-enable profiling stack switch test
Removes the warning on platforms where CFI_NORETURN appears likely to be
sufficient alone for this to work (from observation in gdb/lldb) and
re-enables the test on all platforms so we can see if more work here is
needed at all (e.g. similar annotations in jl_setjmp/jl_longjmp).

Refs #43124