LinearAlgebra: return destination in `setindex!` (#55544)
Currently, in LinearAlgebra, `setindex!` occasionally returns the value,
and at other times the destination array (or its parent). This PR
consistently returns the destination in `setindex!`, which matches the
behavior for `Array`s.

Note that this does not change the behavior of `A[i,j] = v`, which still
returns `v`. This only changes `setindex!`.