math: specialize `sinpi`, `cospi`, `tanpi`
For `sinpi` and `tanpi`, this loses the distinction between the signed
zeros of the previous implementation. That said, as far as I
understand, the distinction was nonsensical, because for both the sine
and the tangens each zero is a zero-crossing (the graph intersects the
x-axis, as opposed to just touching the x-axis).

So, as far as I understand, this change should be fine.