[LinearAlgebra] Initialise number of BLAS threads with `jl_effective_threads`
This is a safer estimate than `Sys.CPU_THREADS` to avoid oversubscribing the
machine when running distributed applications, or when the Julia process is
constrained by external controls (`taskset`, `cgroups`, etc.).

This is also the same function used to determine the number of Julia threads
when starting Julia with `julia -t auto`.