Update merged profile if any changes to raw profile
This addresses the review comment in the original pr, We should also have $(PROFILE_FILE) depend on the result of $(wildcard $(PROFILE_DIR)/*.profraw) so that we automatically rebuild the merged profile data if new profile files have been added.
If stage1 is rebuilt due to profile_file then profile_file will trigger and it no longer uses PROFRAW_FILES (which is defined at parse time of the Makefile) so will use the newest raw files. And if stage1 is not rebuilt but any raw files changes then again the merged profile is built correctly.