prevent loading other extensions when precompiling an extension
The current way of loading extensions when precompiling an extension very easily leads to cycles. For example, if you have more than one extension and you happen to transitively depend on the triggers of one of your extensions you will immidiately hit a cycle.

The test added here checks this scenario and we can now precompile and load it without any warnings or issues.