Call `pthread_attr_init` in `jl_init_stack_limits`
PR 55515 seems to have introduced segfaults on FreeBSD during the
`atexit` tests. Prior to that PR, we had been calling
`pthread_attr_init` in `jl_init_stack_limits` on FreeBSD. According to
the manual page for `pthread_attr_get_np`, it is "HIGHLY RECOMMENDED" to
use `pthread_attr_init` to allocate storage. Might as well put it back
then, and hopefully it'll fix the segfaults.