Try to remove likely-unused codepath in codegen for :method (#55532)
This codepath is odd. It derives a method name from the slotname
metadata and then does an implicit assignment to that slot. Worse, as
the comment indicates, the codepath is missing from the interpreter,
which will crash if it were to ever encounter such a piece of code. I
suspect this pattern is unused - I accidentally broke it badly in (the
as of yet unmerged PR) #54788 and neither tests nor pkgeval noticed. So
let's try removing it on master. If it turns out something does depend
on it, we can go the opposite way and implement it properly in all the
places that look at :method.